pyproject.toml
[project]
name = "cnn-vs-mlp"
version = "0.1.0"
description = "Compare CNN vs MLP on CIFAR-10"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "torch==2.2.2",
  "torchvision==0.17.2",
  "numpy==1.26.4",
  "scikit-learn>=1.5",
  "matplotlib>=3.8",
  "PyYAML>=6.0.1",
  "tqdm>=4.66",
]

[tool.uv]
required-environments = ["sys_platform == 'darwin' and platform_machine == 'x86_64'"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
